backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
redis>=5.0.0
rq>=1.15.0
aiofiles>=23.0.0
